CourseIO now works
Course renamed to terrain
Course is now a set of terrains

Terrain splitted into Terrain and TerrainModel to be able to save only the attributes of the terrain with an object stream.

SubMenu class created to have a uniform pattern of title and return buttons for all submenus.
TerrainCreatorMenu updated.

EditorMenu to hold all terrain and course creating options.

+++ core/src/com/golf2k18/objects/Ball.java
@@ -119,11 +119,11 @@ public class Ball
                 VertexAttributes.Usage.Position | VertexAttributes.Usage.Normal | VertexAttributes.Usage.TextureCoordinates
         );
 
-        model = new ModelInstance(tempSphere,(float)(position.x),(float)(position.y),(float)(position.z+(DIAMETER/2)));
+        model = new ModelInstance(tempSphere,position.x,position.y,position.z+(DIAMETER/2));
     }
 
     private void updateInstance(){
-        model.transform.setTranslation((float)(position.x),(float)(position.y),(float)(position.z+(DIAMETER/2)));
+        model.transform.setTranslation(position.x,position.y,position.z+(DIAMETER/2));
     }
 
     public void stopped(boolean stopped){
